refactor!: move generate_amplitude_model to amplitude
