+++ expertsystem/amplitude/__init__.py
@@ -11,10 +11,39 @@ possible.
 
 
 __all__ = [
+    # Sub-modules
     "canonical_decay",
     "helicity_decay",
     "model",
+    # User Interface
+    "generate_amplitude_model",
 ]
 
 
+from expertsystem.reaction.solving import Result
+
 from . import canonical_decay, helicity_decay, model
+from .canonical_decay import CanonicalAmplitudeGenerator
+from .helicity_decay import HelicityAmplitudeGenerator
+from .model import AmplitudeModel
+
+
+def generate_amplitude_model(result: Result) -> AmplitudeModel:
+    """Generate an amplitude model from a generated `.Result`.
+
+    The type of amplitude model (`.HelicityAmplitudeGenerator` or
+    `.CanonicalAmplitudeGenerator`) is determined from the
+    `.Result.formalism_type`.
+    """
+    formalism_type = result.formalism_type
+    if formalism_type is None:
+        raise ValueError(f"Result does not have a formalism type:\n{result}")
+    if formalism_type == "helicity":
+        amplitude_generator = HelicityAmplitudeGenerator()
+    elif formalism_type in ["canonical-helicity", "canonical"]:
+        amplitude_generator = CanonicalAmplitudeGenerator()
+    else:
+        raise NotImplementedError(
+            f'No amplitude generator for formalism type "{formalism_type}"'
+        )
+    return amplitude_generator.generate(result.solutions)
