+++ expertsystem/ui/__init__.py
@@ -6,33 +6,8 @@ of the `expertsystem`. See the :doc:`/usage/quickstart`.
 
 from typing import List, Set
 
-from expertsystem.amplitude.canonical_decay import CanonicalAmplitudeGenerator
-from expertsystem.amplitude.helicity_decay import HelicityAmplitudeGenerator
-from expertsystem.amplitude.model import AmplitudeModel
 from expertsystem.particle import Particle
-from expertsystem.reaction.solving import Result
 from expertsystem.reaction.topology import StateTransitionGraph
-
-
-def generate_amplitude_model(result: Result) -> AmplitudeModel:
-    """Generate an amplitude model from a generated `.Result`.
-
-    The type of amplitude model (`.HelicityAmplitudeGenerator` or
-    `.CanonicalAmplitudeGenerator`) is determined from the
-    `.Result.formalism_type`.
-    """
-    formalism_type = result.formalism_type
-    if formalism_type is None:
-        raise ValueError(f"Result does not have a formalism type:\n{result}")
-    if formalism_type == "helicity":
-        amplitude_generator = HelicityAmplitudeGenerator()
-    elif formalism_type in ["canonical-helicity", "canonical"]:
-        amplitude_generator = CanonicalAmplitudeGenerator()
-    else:
-        raise NotImplementedError(
-            f'No amplitude generator for formalism type "{formalism_type}"'
-        )
-    return amplitude_generator.generate(result.solutions)
 
 
 def get_intermediate_state_names(
